Fix real var state allocation and enable LongRealVar
AddRealVar picked a variable type in a switch expression, allocating its
state slots, then threw that result away in an if/else below and
allocated a second time. Every real var therefore claimed twice the
state it used, and the LongRealVar branch was unreachable.

Removing the dead branch alone would have introduced a worse bug: the
switch allocated LongRealVar with AddInt() (one word), but LongRealVar
reads and writes through GetLong/SetLong (two words), so it would have
written over the neighbouring variable's slots. It now allocates with
AddLong().

LongRealVar.Initialise also cast its bounds to int, truncating exactly
the values the long backing exists to hold.

Adds ModelBuilderTests covering slot counts for both backings, type
selection at the int/long boundary, bounds surviving Initialise, and
adjacent real vars not overlapping in state. Five of the six fail
against the previous code.

+++ CSPSolver/Variable/LongRealVar.cs
@@ -33,8 +33,8 @@ namespace CSPSolver.Variable
 
         public void Initialise(IState state)
         {
-            state.SetLong(MaxStateRef, (int)Round(Max / Epsilon));
-            state.SetLong(MinStateRef, (int)Round(Min / Epsilon));
+            state.SetLong(MaxStateRef, (long)Round(Max / Epsilon));
+            state.SetLong(MinStateRef, (long)Round(Min / Epsilon));
         }
 
         public bool IsEmpty(IState state) => state.GetLong(MaxStateRef) < state.GetLong(MinStateRef);
